# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.DiagnosticsResponse:
    properties:
      disk:
        $ref: '#/definitions/handlers.DiskDiagnostics'
      hlsCache:
        $ref: '#/definitions/handlers.HlsCacheDiagnostics'
      pieceCache:
        $ref: '#/definitions/handlers.PieceCacheDiagnostics'
      runtime:
        $ref: '#/definitions/handlers.RuntimeDiagnostics'
      sessions:
        $ref: '#/definitions/handlers.SessionDiagnostics'
      thumbCache:
        $ref: '#/definitions/handlers.ThumbCacheDiagnostics'
      torrents:
        items:
          $ref: '#/definitions/handlers.DiagnosticsTorrentEntry'
        type: array
      transcoders:
        $ref: '#/definitions/handlers.TranscoderDiagnostics'
    type: object
  handlers.DiagnosticsTorrentEntry:
    properties:
      activePeers:
        type: integer
      cacheBytes:
        type: integer
      diskBytes:
        type: integer
      downloadMode:
        type: string
      hash:
        type: string
      name:
        type: string
      numPieces:
        type: integer
      piecesComplete:
        type: integer
      pinned:
        type: boolean
      watchers:
        type: integer
    type: object
  handlers.DiskDiagnostics:
    properties:
      dir:
        type: string
      free:
        type: integer
      used:
        type: integer
    type: object
...
  handlers.HlsCacheDiagnostics:
    properties:
      bytes:
        type: integer
      count:
        type: integer
      max:
        type: integer
    type: object
...
  handlers.PieceCacheDiagnostics:
    properties:
      capacity:
        type: integer
      filled:
        type: integer
    type: object
...
  handlers.RuntimeDiagnostics:
    properties:
      gcCPUFraction:
        type: number
      heapAlloc:
        type: integer
      heapSys:
        type: integer
      memLimitBytes:
        type: integer
      numGC:
        type: integer
      numGoroutine:
        type: integer
      stackSys:
        type: integer
      sys:
        type: integer
      uptimeSec:
        type: integer
    type: object
...
  handlers.SessionDiagnostics:
    properties:
      active:
        type: integer
      maxStreams:
        type: integer
      perStreamBytes:
        type: integer
      streaming:
        type: integer
    type: object
...
  handlers.ThumbCacheDiagnostics:
    properties:
      bytes:
        type: integer
      count:
        type: integer
      max:
        type: integer
    type: object
...
  handlers.TranscoderDiagnostics:
    properties:
      count:
        type: integer
      estBytes:
        type: integer
      max:
        type: integer
    type: object
info:
  contact: {}
  description: |-
    Standalone BitTorrent streaming engine for Potok: torrent metadata, direct streaming, HLS, subtitles, thumbnails, and an optional management web UI.
    The plugin/streaming routes are open (players hit them without credentials); the /api/manage/* contour is mounted only with TORRENTGO_ENABLE_WEBUI=true and protected by HTTP Basic auth. Errors are text/plain (via http.Error) unless noted otherwise.
  title: Potok TorrentGo API
  version: "2.0"
paths:
  /api/manage/diagnostics:
    get:
      description: A full accounting of where RAM (and disk) goes — the Go runtime
        heap, the global piece cache (with per-torrent breakdown), the HLS segment
        cache, the thumbnail cache, live AAC transcoders, playback sessions, and disk
        usage. Everything the operator needs to see "what is eating memory" in one
        place.
      operationId: getDiagnostics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.DiagnosticsResponse'
        "401":
          description: unauthorized (BasicAuth)
          schema:
            type: string
      security:
      - BasicAuth: []
      summary: Full memory/disk diagnostics
      tags:
      - Management
      x-scalar-ignore: true
...
  /api/torrents/{hash}/files/{fileIndex}/subtitle-file:
    get:
      description: Reads a WHOLE sidecar subtitle file (a separate file in the torrent,
        "ext" releases) by its 1-based torrent index and converts SRT to WebVTT (ASS/SSA/VTT
        pass through unchanged). Unlike /subtitles/{trackIndex}, no embedded-track
        demux is involved; `?start=` is ignored (sub files are tiny, the player's
        windowed feeder dedupes cues).
      operationId: getExternalSubtitleFile
      parameters:
      - description: Infohash (40-char hex)
        in: path
        name: hash
        required: true
        type: string
      - description: Torrent index of the subtitle file itself (1-based)
        in: path
        name: fileIndex
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "200":
          description: Subtitle body (text/vtt, text/x-ssa or text/plain)
          schema:
            type: string
        "500":
          description: subtitle file unavailable / read failed
          schema:
            type: string
      security: []
      summary: Serve an external (sidecar) subtitle file
      tags:
      - Media
...
schemes:
- http
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"